+++ Classes/Authentication/FrontendAuthenticationService.php
@@ -2,20 +2,6 @@
 
 declare(strict_types=1);
 
-/*
- * This file is part of the OAuth2 Client extension for TYPO3
- * - (c) 2021 waldhacker UG (haftungsbeschränkt)
- *
- * It is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License, either version 2
- * of the License, or any later version.
- *
- * For the full copyright and license information, please read the
- * LICENSE.txt file that was distributed with this source code.
- *
- * The TYPO3 project - inspiring people to share!
- */
-
 namespace Waldhacker\Oauth2Client\Authentication;
 
 use Doctrine\DBAL\Exception;
@@ -103,7 +89,7 @@ class FrontendAuthenticationService extends AbstractAuthenticationService
         return -100;
     }
 
-    public function processLoginData(array &$loginData, string $_): bool
+    public function processLoginData(array &$loginData): bool
     {
         $loginData['uname'] = $loginData['uname'] ?? '';
         $loginData['uident'] = $loginData['uident'] ?? '';