+++ Classes/Controller/Backend/ManageProvidersController.php
@@ -2,20 +2,6 @@
 
 declare(strict_types=1);
 
-/*
- * This file is part of the OAuth2 Client extension for TYPO3
- * - (c) 2021 waldhacker UG (haftungsbeschränkt)
- *
- * It is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License, either version 2
- * of the License, or any later version.
- *
- * For the full copyright and license information, please read the
- * LICENSE.txt file that was distributed with this source code.
- *
- * The TYPO3 project - inspiring people to share!
- */
-
 namespace Waldhacker\Oauth2Client\Controller\Backend;
 
 use Doctrine\DBAL\Exception;
@@ -23,27 +9,25 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use TYPO3\CMS\Backend\Routing\Exception\RouteNotFoundException;
 use TYPO3\CMS\Backend\Routing\UriBuilder;
-use TYPO3\CMS\Backend\Template\Components\ButtonBar;
 use TYPO3\CMS\Backend\Template\ModuleTemplate;
 use TYPO3\CMS\Backend\Template\ModuleTemplateFactory;
 use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Context\Exception\AspectNotFoundException;
-use TYPO3\CMS\Core\Imaging\Icon;
 use TYPO3\CMS\Core\Imaging\IconFactory;
-use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Imaging\IconSize;
+use TYPO3\CMS\Core\Localization\LanguageService;
 use Waldhacker\Oauth2Client\Repository\BackendUserRepository;
 use Waldhacker\Oauth2Client\Service\Oauth2ProviderManager;
 
-class ManageProvidersController extends AbstractBackendController
+readonly class ManageProvidersController
 {
     public function __construct(
-        private readonly Oauth2ProviderManager $oauth2ProviderManager,
-        private readonly BackendUserRepository $backendUserRepository,
-        private readonly UriBuilder $uriBuilder,
-        private readonly IconFactory $iconFactory,
-        private readonly ModuleTemplateFactory $moduleTemplateFactory,
-        private readonly Context $context,
+        private Oauth2ProviderManager $oauth2ProviderManager,
+        private BackendUserRepository $backendUserRepository,
+        private UriBuilder $uriBuilder,
+        private IconFactory $iconFactory,
+        private ModuleTemplateFactory $moduleTemplateFactory,
+        private Context $context,
     ) {
     }
 
@@ -72,44 +56,17 @@ class ManageProvidersController extends AbstractBackendController
         $languageFile = 'LLL:EXT:core/Resources/Private/Language/locallang_core.xlf:';
         $buttonBar = $moduleTemplate->getDocHeaderComponent()->getButtonBar();
 
-        if (($returnUrl = $this->getReturnUrl($request)) !== '') {
-            $button = $buttonBar
-                ->makeLinkButton()
-                ->setHref($returnUrl)
-                ->setIcon($this->iconFactory->getIcon('actions-view-go-back', Icon::SIZE_SMALL))
-                ->setTitle($this->getLanguageService()->sL($languageFile . 'labels.goBack'))
-                ->setShowLabelText(true);
-            $buttonBar->addButton($button);
-        }
-
-        $reloadButton = $buttonBar
+        $button = $buttonBar
             ->makeLinkButton()
-            ->setHref($request->getAttribute('normalizedParams')->getRequestUri())
-            ->setTitle($this->getLanguageService()->sL($languageFile . 'labels.reload'))
-            ->setIcon($this->iconFactory->getIcon('actions-refresh', Icon::SIZE_SMALL));
-        $buttonBar->addButton($reloadButton, ButtonBar::BUTTON_POSITION_RIGHT);
+            ->setHref((string)$this->uriBuilder->buildUriFromRoute('user_setup'))
+            ->setIcon($this->iconFactory->getIcon('actions-view-go-back', IconSize::SMALL))
+            ->setTitle($this->getLanguageService()->sL($languageFile . 'labels.goBack'))
+            ->setShowLabelText(true);
+        $buttonBar->addButton($button);
     }
 
-    /**
-     * @throws RouteNotFoundException
-     */
-    private function getReturnUrl(ServerRequestInterface $request): string
+    protected function getLanguageService(): LanguageService
     {
-        $queryParams = $request->getQueryParams();
-        $parsedBody = $request->getParsedBody();
-        if (isset($queryParams['returnUrl'])) {
-            $returnUrl = $queryParams['returnUrl'];
-        } elseif (is_array($parsedBody) && isset($parsedBody['returnUrl'])) {
-            $returnUrl = $parsedBody['returnUrl'];
-        } else {
-            $returnUrl = '';
-        }
-        $returnUrl = GeneralUtility::sanitizeLocalUrl($returnUrl);
-
-        if ($returnUrl === '' && ExtensionManagementUtility::isLoaded('setup')) {
-            $returnUrl = (string)$this->uriBuilder->buildUriFromRoute('user_setup');
-        }
-
-        return $returnUrl;
+        return $GLOBALS['LANG'];
     }
 }